Rename some internal modules (#68)
