+++ geofileops/geofile.py
@@ -6,4 +6,4 @@ Only for backwards compatibility.
 import warnings
 warnings.warn("using from geofileops import geofile is deprecated, please just use import geofileops", FutureWarning)
 
-from .file import *
+from .fileops import *
