+++ geofileops/geofileops.py
@@ -6,4 +6,4 @@ Only for backwards compatibility.
 import warnings
 warnings.warn("using from geofileops import geofileops is deprecated, please just use import geofileops", FutureWarning)
 
-from .fileops import *
+from .geoops import *
